Improve structure
