+++ src/main/kotlin/no/nav/tpts/mottak/soknad/soknadList/Soknad.kt
@@ -1,4 +1,4 @@
-package no.nav.tpts.mottak.soknad
+package no.nav.tpts.mottak.soknad.soknadList
 
 import kotlinx.serialization.Serializable
 import no.nav.tpts.mottak.databind.LocalDateSerializer
@@ -6,20 +6,13 @@ import no.nav.tpts.mottak.databind.LocalDateTimeSerializer
 import java.time.LocalDate
 import java.time.LocalDateTime
 
-enum class SoknadStatus {
-    AVSLAG,
-    IKKE_BEHANDLET,
-    BEHANDLES
-}
-
 @Serializable
 data class Soknad(
     val id: String,
     val fornavn: String,
     val etternavn: String,
-    val identer: List<String>,
-    val soknadStatus: SoknadStatus = SoknadStatus.IKKE_BEHANDLET,
-    @Serializable(with = LocalDateTimeSerializer::class) val opprettetDato: LocalDateTime,
+    val ident: String,
+    @Serializable(with = LocalDateTimeSerializer::class) val opprettet: LocalDateTime,
     @Serializable(with = LocalDateSerializer::class) val brukerStartDato: LocalDate?,
     @Serializable(with = LocalDateSerializer::class) val brukerSluttDato: LocalDate?
 )
