+++ src/test/resources/soknad.json
@@ -1,6 +1,6 @@
 {
   "fornavn": "Sigurd",
-  "opprettetDato": "+999999999-12-31T23:59:59.999999999",
+  "opprettet": "+999999999-12-31T23:59:59.999999999",
   "brukerStartDato": null,
   "brukerSluttDato": null,
   "fnr": "12121221212"
